修正对 _open_osfhandle 参数的转换
在 MSYS2/MinGW64 下编译报错

+++ src/egegapi.cpp
@@ -82,13 +82,13 @@ BOOL init_console(){
 		DrawMenuBar(hConsoleWnd);
 	}
 
-	hCrt=_open_osfhandle((long)hOutput,_O_TEXT);
+	hCrt=_open_osfhandle((intptr_t)hOutput,_O_TEXT);
 	hf=_fdopen( hCrt, "w" );
 	fOldOut=*stdout;
 	*stdout=*hf;
 	setvbuf( stdout, NULL, _IONBF, 0 );
 
-	hCrt=_open_osfhandle((long)hInput,_O_TEXT);
+	hCrt=_open_osfhandle((intptr_t)hInput,_O_TEXT);
 	hf=_fdopen(hCrt, "r");
 	fOldIn=*stdin;
 	*stdin=*hf;
